♻️ Convert `:feature:alarm` to KMP
One more module converted to KMP, allowing to be reused with the
`:feature:task` module. Several key parts are platform-dependent will
are now on its own source. Future commits will connect and test
everything together.
