+++ features/alarm/src/commonMain/kotlin/com/escodro/alarm/permission/PermissionChecker.kt
@@ -1,0 +1,23 @@
+package com.escodro.alarm.permission
+
+/**
+ * Represents the Android permission checker.
+ */
+internal interface PermissionChecker {
+
+    /**
+     * Checks if the given permission was granted by the user.
+     *
+     * @param permission the OS permission
+     *
+     * @return `true` if the permission is granted, `false` otherwise
+     */
+    fun checkPermission(permission: String): Boolean
+
+    /**
+     * Checks if the special schedule exact alarms permission was granted by the user.
+     *
+     * @return `true` if the permission is granted, `false` otherwise
+     */
+    fun canScheduleExactAlarms(): Boolean
+}
