+++ features/alarm/src/commonMain/kotlin/com/escodro/alarm/permission/SdkVersion.kt
@@ -1,0 +1,12 @@
+package com.escodro.alarm.permission
+
+/**
+ * Represents the current OS SDK version.
+ */
+internal interface SdkVersion {
+
+    /**
+     * The current OS SDK version.
+     */
+    val currentVersion: Int
+}
