+++ features/alarm/src/iosMain/kotlin/com/escodro/alarm/di/PlatformAlarmModule.kt
@@ -1,0 +1,8 @@
+package com.escodro.alarm.di
+
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+actual val platformAlarmModule: Module = module {
+    // TODO
+}
