--- features/alarm/src/main/java/com/escodro/alarm/permission/AndroidVersion.kt
@@ -1,12 +1,0 @@
-package com.escodro.alarm.permission
-
-/**
- * Represents the current Android SDK version.
- */
-internal interface AndroidVersion {
-
-    /**
-     * The current Android SDK version.
-     */
-    val currentVersion: Int
-}
