--- features/alarm/src/main/res/values/strings.xml
@@ -1,12 +1,0 @@
-<resources>
-    <!-- Notification app name -->
-    <string name="content_app_name" translatable="false">Alkaa</string>
-
-    <!-- Task notification channel -->
-    <string name="channel_task_name">Task alarms</string>
-    <string name="channel_task_description">Notifications for reminders</string>
-
-    <!-- Task notification -->
-    <string name="notification_action_completed">Done</string>
-    <string name="notification_action_snooze">In 15 min</string>
-</resources>
