+++ resources/src/commonMain/resources/MR/pt-br/strings.xml
@@ -103,4 +103,11 @@
     <string name="tracker_header_empty">Você ainda não completou nenhuma tarefa nos últimos 30 dias</string>
     <string name="tracker_header_error">Ops… Isso é embaraçoso…</string>
     <string name="tracker_cd_error">Algo deu errado</string>
+
+    <!-- Alarm -->
+    <string name="channel_task_name">Alarme de tarefas</string>
+    <string name="channel_task_description">Notificações para lembretes</string>
+
+    <string name="notification_action_completed">Concluído</string>
+    <string name="notification_action_snooze">Daqui 15 min</string>
 </resources>
